+++ docker/run_tests.sh
@@ -18,7 +18,6 @@ rm "$INTEGRATION_TESTS"*
 
 echo -e "${GREEN}*** Inject system allocator ***${NC}"
 sed -i "/\/\/>alloc_system/ c\#![feature(alloc_system, global_allocator, allocator_api)]\nextern crate alloc_system;\nuse alloc_system::System;\n\#[global_allocator]\nstatic A: System = System;\n" src/lib.rs
-cat src/lib.rs
 
 echo -e "${GREEN}*** Build tests ***${NC}"
 cargo test --no-run